+++ src/Http/Controllers/SocialiteAuthController.php
@@ -9,10 +9,10 @@ use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\URL;
-use InternetGuru\LaravelUser\Enums\Provider;
 use InternetGuru\LaravelUser\Enums\ProviderAction;
 use InternetGuru\LaravelUser\Exceptions\AuthCheckException;
 use Laravel\Socialite\Facades\Socialite;
+use InternetGuru\LaravelUser\Models\Socialite as SocialiteModel;
 
 class SocialiteAuthController extends Controller
 {
@@ -36,7 +36,7 @@ class SocialiteAuthController extends Controller
     public function handleProviderAction(string $provider, string $action, Request $request): RedirectResponse
     {
         try {
-            $provider = Provider::from($provider);
+            $provider = SocialiteModel::providers()::from($provider);
             $action = ProviderAction::from($action);
 
             // switch the actions
@@ -87,7 +87,7 @@ class SocialiteAuthController extends Controller
     public function handleProviderCallback(string $provider, string $action): RedirectResponse
     {
         try {
-            $provider = Provider::from($provider);
+            $provider = SocialiteModel::providers()::from($provider);
             $action = ProviderAction::from($action);
 
             try {