+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Changed
+
+- Make Provider enum customizable.
+
 ## [2.3.4] - 2025-09-02
 
 ### Fixed
@@ -493,6 +499,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[Unreleased]: https://https://github.com/internetguru/laravel-user/compare/staging...dev
 [2.3.4]: https://https://github.com/internetguru/laravel-user/compare/v2.3.3...v2.3.4
 [2.3.3]: https://https://github.com/internetguru/laravel-user/compare/v2.3.2...v2.3.3
 [2.3.2]: https://https://github.com/internetguru/laravel-user/compare/v2.3.1...v2.3.2